Create NUMBER GUESSING GAME
This is a number guessing game.
Initially user must give a range in which a random number must be generated. 
And user must guess the random number generated.
Until user guesses the correct answer the continues.
It shows how many chances does the user took to guess the randomly generated number.
After completion of a game , the user is asked to stop or continue the game.